Rename parseSlots to parseSlotsAndClasses and improve it to generate classes
